+++ include/iot/data.h
@@ -434,7 +434,7 @@ extern iot_data_t * iot_data_alloc_string (const char * val, iot_data_ownership_
  *
  * @param data       Address of static storage for data
  * @param str        Constant string value.
- * @return           Pointer to the allocated data (same address as static storge)
+ * @return           Pointer to the allocated data (same address as the static storge)
  */
 extern iot_data_t * iot_data_alloc_const_string (iot_data_static_t * data, const char * str);
 
@@ -461,6 +461,18 @@ extern iot_data_t * iot_data_alloc_string_fmt (const char * format, ...);
  * @return           Pointer to the allocated data
  */
 extern iot_data_t * iot_data_alloc_pointer (void * ptr, iot_data_free_fn free_fn);
+
+/**
+ * @brief Allocate constant pointer data
+ *
+ * The function to allocate data for a constant pointer, using fixed static storage, so need
+ * not be deleted.
+ *
+ * @param data       Address of static storage for data
+ * @param ptr        Constant pointer value.
+ * @return           Pointer to the allocated data (same address as the static storge)
+ */
+extern iot_data_t * iot_data_alloc_const_pointer (iot_data_static_t * data, const void * ptr);
 
 /**
  * @brief Allocate data for a list
@@ -1394,6 +1406,38 @@ extern const void * iot_data_array_iter_value (const iot_data_array_iter_t * ite
  * @param iter  Iterator to initialise
  */
 extern void iot_data_map_iter (const iot_data_t * map, iot_data_map_iter_t * iter);
+
+/**
+ * @brief Return first element in a map or NULL if map empty
+ *
+ * @param  map   Input map
+ * @return       First element in the map
+ */
+extern const iot_data_t * iot_data_map_start (iot_data_t * map);
+
+/**
+ * @brief Return the value of the first pointer element in a map or NULL if map empty
+ *
+ * @param  map   Input map
+ * @return       Pointer from the first element in the map
+ */
+extern const void * iot_data_map_start_pointer (iot_data_t * map);
+
+/**
+ * @brief Return last element in a map or NULL if map empty
+ *
+ * @param  map   Input map
+ * @return       Last element in the map
+ */
+extern const iot_data_t * iot_data_map_end (iot_data_t * map);
+
+/**
+ * @brief Return the value of the last pointer element in a map or NULL if map empty
+ *
+ * @param  map   Input map
+ * @return       Pointer from the last element in the map
+ */
+extern const void * iot_data_map_end_pointer (iot_data_t * map);
 
 /**
  * @brief Update the iterator to point to the next element within a map