+++ RELEASE_NOTES.md
@@ -225,10 +225,18 @@
 * `iot_data_transform`
 * `iot_data_vector_element_count`
 * `iot_data_vector_dimensions`
-
+* `iot_data_map_start`
+* `iot_data_map_start_pointer`
+* `iot_data_map_end`
+* `iot_data_map_end_pointer`
+* 
 - Function added to return size of encapsulated C data type
 
 * `iot_data_type_size`
+
+- Scheduler implementation scalability rework, plus function added to return schedule id:
+
+* `iot_schedule_id`
 
 - Map data type updated to support all key types
 - Fedora support updated for Fedora 35